Aligning documentation to the new folder structure. Including the new AscEditor class on the description
