+++ PyLTSpice/sim/tookit/montecarlo.py
@@ -1,0 +1,19 @@
+#!/usr/bin/env python
+# coding=utf-8
+
+# -------------------------------------------------------------------------------
+#    ____        _   _____ ____        _
+#   |  _ \ _   _| | |_   _/ ___| _ __ (_) ___ ___
+#   | |_) | | | | |   | | \___ \| '_ \| |/ __/ _ \
+#   |  __/| |_| | |___| |  ___) | |_) | | (_|  __/
+#   |_|    \__, |_____|_| |____/| .__/|_|\___\___|
+#          |___/                |_|
+#
+# Name:        montecarlo.py
+# Purpose:     Classes to automate Monte-Carlo simulations
+#
+# Author:      Nuno Brum (nuno.brum@gmail.com)
+#
+# Created:     10-08-2023
+# Licence:     refer to the LICENSE file
+# -------------------------------------------------------------------------------
