--- doc/classes/SpiceCircuit.rst
@@ -1,7 +1,0 @@
-SpiceCircuit
-============
-
-.. autoclass:: PyLTSpice.SpiceEditor.SpiceCircuit
-   :members:
-   :undoc-members:
-   :show-inheritance:
