--- doc/classes/SpiceEditor.rst
@@ -1,7 +1,0 @@
-SpiceEditor
-===========
-
-.. autoclass:: PyLTSpice.SpiceEditor.SpiceEditor
-   :members:
-   :undoc-members:
-   :show-inheritance:
