+++ doc/classes/asc_editor.rst
@@ -1,0 +1,9 @@
+AscEditor
+============
+
+Class used for manipulating LTSpice asc files.
+
+.. autoclass:: PyLTSpice.editor.asc_editor.AscEditor
+   :members:
+   :undoc-members:
+   :show-inheritance:
