+++ doc/classes/raw_classes.rst
@@ -1,0 +1,13 @@
+========================
+Raw and Log File Classes
+========================
+
+.. toctree::
+   :maxdepth: 4
+   
+   raw_read
+   trace
+   logreader
+   raw_write
+   write_trace
+
