+++ doc/classes/sim_client.rst
@@ -1,5 +1,6 @@
-SimClient (Under development)
-=============================
+=========
+SimClient
+=========
 
 .. autoclass:: PyLTSpice.client_server.sim_client.SimClient
    :members:
