+++ doc/classes/simulation_classes.rst
@@ -1,16 +1,11 @@
-==============
-Python Classes
-==============
-
+==================
+Simulation Classes
+==================
 .. toctree::
    :maxdepth: 4
-
-   spice_editor
-   spice_circuit
+   
    sim_runner
    sim_simulator
-   raw_read
-   raw_write
    sim_client
    sim_server
