+++ doc/classes/spice_circuit.rst
@@ -1,7 +1,7 @@
 SpiceCircuit
 ============
 
-.. autoclass:: PyLTSpice.sim.spice_editor.SpiceCircuit
+.. autoclass:: PyLTSpice.editor.spice_editor.SpiceCircuit
    :members:
    :undoc-members:
    :show-inheritance:
