+++ doc/classes/spice_editor.rst
@@ -3,7 +3,7 @@ SpiceEditor
 
 Class used for manipulating SPICE netlists. Inherits from SpiceCircuit.
 
-.. autoclass:: PyLTSpice.sim.spice_editor.SpiceEditor
+.. autoclass:: PyLTSpice.editor.spice_editor.SpiceEditor
    :members:
    :undoc-members:
    :show-inheritance:
