+++ doc/classes/trace.rst
@@ -1,7 +1,12 @@
 RawRead Trace
 =============
 
-.. autoclass:: PyLTSpice.raw_classes.Trace
+.. autoclass:: PyLTSpice.raw.raw_classes.Axis
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. autoclass:: PyLTSpice.raw.raw_classes.TraceRead
    :members:
    :undoc-members:
    :show-inheritance:
