+++ doc/classes/write_trace.rst
@@ -1,7 +1,7 @@
 RawWrite Trace
 ==============
 
-.. autoclass:: PyLTSpice.raw_write.Trace
+.. autoclass:: PyLTSpice.raw.raw_write.Trace
    :members:
    :undoc-members:
    :show-inheritance:
